+++ lib/screens/accounts_screen/accounts_screen_view.dart
@@ -1,16 +1,18 @@
-import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:vault/constants/app_colors.dart';
-import 'package:vault/constants/app_images.dart';
-import 'package:vault/screens/account_details_screen/account_details_screen_view.dart';
 import 'package:vault/screens/accounts_screen/accounts_screen_controller.dart';
-import 'package:vault/screens/add_account_screen/add_account_screen_view.dart';
-import 'package:vault/services/account_service.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_sizer/responsive_sizer.dart';
-import 'package:cached_network_image/cached_network_image.dart';
+import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:vault/screens/accounts_screen/widgets/custom_account.dart';
+import 'package:vault/screens/home_screen/home_screen_controller.dart';
 
 class AccountsScreenView extends StatefulWidget {
-  const AccountsScreenView({Key? key}) : super(key: key);
+  const AccountsScreenView({
+    Key? key,
+    required this.accountsScreenController,
+  }) : super(key: key);
+
+  final HomeScreenController accountsScreenController;
 
   @override
   State<AccountsScreenView> createState() => _AccountsScreenViewState();
@@ -26,7 +28,7 @@ class _AccountsScreenViewState extends State<AccountsScreenView> {
       context: context,
     );
 
-    controller.queryAccountItem();
+    controller.getAccountsData();
     super.initState();
   }
 
@@ -34,10 +36,16 @@ class _AccountsScreenViewState extends State<AccountsScreenView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        leading: IconButton(
+          onPressed: () => Navigator.pop(context),
+          icon: const Icon(
+            Icons.arrow_back_ios,
+          ),
+        ),
         title: Text(
-          "THE VAULT",
+          "ACCOUNTS",
           style: TextStyle(
-            fontSize: 20.sp,
+            fontSize: 18.5.sp,
             fontWeight: FontWeight.w500,
           ),
         ),
@@ -45,79 +53,23 @@ class _AccountsScreenViewState extends State<AccountsScreenView> {
       ),
       body: Padding(
         padding: EdgeInsets.symmetric(vertical: 2.h, horizontal: 5.w),
-        child: accounts.isEmpty
+        child: controller.isLoading
             ? Center(
-                child: Text(
-                  "There are no accounts found.",
-                  style: TextStyle(
-                    fontSize: 17.5.sp,
-                  ),
+                child: SpinKitFadingCircle(
+                  color: color02,
+                  size: 27.sp,
+                  duration: const Duration(milliseconds: 1400),
                 ),
               )
             : ListView.builder(
-                physics: const ClampingScrollPhysics(),
-                padding: EdgeInsets.zero,
-                itemCount: accounts.length,
+                itemCount: controller.accounts.length,
                 itemBuilder: (context, index) {
-                  return GestureDetector(
-                    onTap: () => Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => AccountDetailsScreenView(
-                          accountItem: accounts[index],
-                          accountsScreenController: controller,
-                        ),
-                      ),
-                    ),
-                    child: Container(
-                      padding: EdgeInsets.symmetric(
-                        vertical: 2.w,
-                        horizontal: 1.w,
-                      ),
-                      margin: EdgeInsets.only(bottom: 3.w),
-                      decoration: BoxDecoration(
-                        color: color06,
-                        border: Border.all(
-                          color: color04,
-                        ),
-                        borderRadius: BorderRadius.circular(7.5),
-                      ),
-                      child: ListTile(
-                        leading: CachedNetworkImage(
-                          imageUrl: accounts[index].accountItemImage,
-                          placeholder: (context, url) => SpinKitFadingCircle(
-                            color: color02,
-                            size: 22.sp,
-                            duration: const Duration(milliseconds: 1400),
-                          ),
-                          errorWidget: (context, url, error) =>
-                              Image.asset(imageLogo),
-                          width: 11.w,
-                          height: 11.w,
-                          fit: BoxFit.contain,
-                        ),
-                        title: Text(
-                          accounts[index].accountItemName,
-                          style: TextStyle(
-                            fontSize: 18.sp,
-                          ),
-                        ),
-                      ),
-                    ),
+                  return CustomAccount(
+                    account: controller.accounts[index],
+                    accountsScreenController: widget.accountsScreenController,
                   );
                 },
               ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () => Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => AddAccountScreenView(
-              accountsScreenController: controller,
-            ),
-          ),
-        ),
-        child: const Icon(Icons.add),
       ),
     );
   }
