+++ lib/screens/accounts_screen/widgets/custom_account_item.dart
@@ -2,11 +2,13 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:vault/constants/app_colors.dart';
 import 'package:vault/constants/app_images.dart';
 import 'package:vault/models/account_item_model.dart';
-import 'package:vault/screens/accounts_screen/accounts_screen_controller.dart';
-import 'package:vault/screens/add_account_details_screen/add_account_details_screen_view.dart';
+// import 'package:vault/screens/account_credentials_screen/account_credentials_screen_controller.dart';
+import 'package:vault/screens/account_credentials_screen/account_credentials_screen_view.dart';
+// import 'package:vault/screens/add_account_details_screen/add_account_details_screen_view.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_sizer/responsive_sizer.dart';
 import 'package:cached_network_image/cached_network_image.dart';
+import 'package:vault/screens/home_screen/home_screen_controller.dart';
 
 class CustomAccountItem extends StatefulWidget {
   const CustomAccountItem({
@@ -16,7 +18,7 @@ class CustomAccountItem extends StatefulWidget {
   }) : super(key: key);
 
   final AccountItem accountItem;
-  final AccountsScreenController accountsScreenController;
+  final HomeScreenController accountsScreenController;
 
   @override
   State<CustomAccountItem> createState() => _CustomAccountItemState();
@@ -28,10 +30,18 @@ class _CustomAccountItemState extends State<CustomAccountItem> {
     return GestureDetector(
       onTap: () => Navigator.push(
         context,
+        // MaterialPageRoute(
+        //   builder: (context) => AddAccountDetailsScreenView(
+        //     accountItem: widget.accountItem,
+        //     accountsScreenController: widget.accountsScreenController,
+        //   ),
+        // ),
         MaterialPageRoute(
-          builder: (context) => AddAccountDetailsScreenView(
+          builder: (context) => AccountCredentialsScreenView(
             accountItem: widget.accountItem,
-            accountsScreenController: widget.accountsScreenController,
+            homeScreenController: widget.accountsScreenController,
+            isAddAccount: true,
+            isUpdateDeleteAccount: false,
           ),
         ),
       ),
