+++ lib/screens/add_account_details_screen/add_account_details_screen_controller.dart
@@ -1,5 +1,5 @@
 import 'package:vault/models/account_item_model.dart';
-import 'package:vault/screens/accounts_screen/accounts_screen_controller.dart';
+import 'package:vault/screens/home_screen/home_screen_controller.dart';
 import 'package:vault/services/account_service.dart';
 import 'package:vault/shared/shared_dialog.dart';
 import 'package:flutter/material.dart';
@@ -8,7 +8,7 @@ import 'package:responsive_sizer/responsive_sizer.dart';
 class AddAccountDetailsScreenController {
   final Function() setstate;
   final BuildContext context;
-  final AccountsScreenController accountsScreenController;
+  final HomeScreenController accountsScreenController;
 
   AddAccountDetailsScreenController({
     required this.setstate,
