+++ lib/screens/home_screen/home_screen_controller.dart
@@ -1,0 +1,17 @@
+import 'package:vault/services/account_service.dart';
+import 'package:flutter/material.dart';
+
+class HomeScreenController {
+  final Function() setstate;
+  final BuildContext context;
+
+  HomeScreenController({
+    required this.setstate,
+    required this.context,
+  });
+
+  Future<void> queryAccountItem() async {
+    await AccountService().queryAccountItem();
+    setstate();
+  }
+}
