+++ lib/screens/set_pin_confirmation_screen/set_pin_confirmation_screen_controller.dart
@@ -1,4 +1,6 @@
-import 'package:vault/screens/accounts_screen/accounts_screen_view.dart';
+import 'package:responsive_sizer/responsive_sizer.dart';
+import 'package:vault/constants/app_colors.dart';
+import 'package:vault/screens/home_screen/home_screen_view.dart';
 import 'package:vault/services/shared_preferences_service.dart';
 import 'package:vault/shared/shared_loading.dart';
 import 'package:vault/shared/shared_snackbar.dart';
@@ -38,7 +40,7 @@ class SetPinConfirmationScreenController {
             Navigator.pushReplacement(
               context,
               MaterialPageRoute(
-                builder: (context) => const AccountsScreenView(),
+                builder: (context) => const HomeScreenView(),
               ),
             );
           },
@@ -55,7 +57,13 @@ class SetPinConfirmationScreenController {
 
             showSharedSnackbar(
               context: context,
-              content: const Text("Pin does not match."),
+              content: Text(
+                "Pin does not match.",
+                style: TextStyle(
+                  color: color04,
+                  fontSize: 17.sp,
+                ),
+              ),
             );
           },
         );
