+++ lib/shared/shared_snackbar.dart
@@ -1,3 +1,4 @@
+import 'package:vault/constants/app_colors.dart';
 import 'package:flutter/material.dart';
 
 showSharedSnackbar({
@@ -7,6 +8,11 @@ showSharedSnackbar({
   return ScaffoldMessenger.of(context).showSnackBar(
     SnackBar(
       content: content,
+      backgroundColor: color02,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      behavior: SnackBarBehavior.floating,
       duration: const Duration(milliseconds: 1000),
     ),
   );
